+++ CHANGELOG.md
@@ -11,11 +11,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Readded Room::unreadMessageCount until NotificationService is updated
+
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+
+- Fixed Room::isUnread calculation
 
 ### Security